requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers
chromadb
transformers
torch
pypdf
streamlit
python-dotenv
pandas
google-genai